phonecalls/requirements.txt
selenium
google-api-python-client
google-cloud-speech
nexmo
requests